requirements.txt
torch
torchvision
torchaudio
streamlit==1.43.1
pillow==11.1.0
numpy==1.26.4
pandas==2.2.1
opencv-python==4.9.0.80
matplotlib==3.8.3
scikit-learn==1.4.1.post1
